remove old commit search path
The old commit search path has been disabled since the v3.33.0 release,
and no longer is functional due to the changes with streaming repos.
This commit removes the old commit search path and all code that becomes
unused because of that.
