+++ internal/search/types.go
@@ -14,7 +14,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/search/query"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 	"github.com/sourcegraph/sourcegraph/internal/types"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 	"github.com/sourcegraph/sourcegraph/schema"
 )
 
@@ -22,38 +21,8 @@ type TypeParameters interface {
 	typeParametersValue()
 }
 
-func (CommitParameters) typeParametersValue()  {}
-func (DiffParameters) typeParametersValue()    {}
 func (SymbolsParameters) typeParametersValue() {}
 func (TextParameters) typeParametersValue()    {}
-
-type CommitParameters struct {
-	RepoRevs           *RepositoryRevisions
-	PatternInfo        *CommitPatternInfo
-	Query              query.Q
-	Diff               bool
-	ExtraMessageValues []string
-}
-
-type DiffParameters struct {
-	Repo    api.RepoName
-	Options git.RawLogDiffSearchOptions
-}
-
-// CommitPatternInfo is the data type that describes the properties of
-// a pattern used for commit search.
-type CommitPatternInfo struct {
-	Pattern         string
-	IsRegExp        bool
-	IsCaseSensitive bool
-	FileMatchLimit  int32
-
-	IncludePatterns []string
-	ExcludePattern  string
-
-	PathPatternsAreRegExps       bool
-	PathPatternsAreCaseSensitive bool
-}
 
 type SymbolsParameters struct {
 	// Repo is the name of the repository to search in.
@@ -194,16 +163,6 @@ type TextParameters struct {
 
 	Zoekt        zoekt.Streamer
 	SearcherURLs *endpoint.Map
-}
-
-// TextParametersForCommitParameters is an intermediate type based on
-// TextParameters that encodes parameters exclusively for a commit search. The
-// commit search internals converts this type to CommitParameters. The
-// commitParameter type definitions will be merged in future.
-type TextParametersForCommitParameters struct {
-	PatternInfo *CommitPatternInfo
-	Repos       []*RepositoryRevisions
-	Query       query.Q
 }
 
 // TextPatternInfo is the struct used by vscode pass on search queries. Keep it in
