+++ internal/vcs/git/commits.go
@@ -1,11 +1,9 @@
 package git
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"os"
 	"strconv"
 	"strings"
@@ -292,82 +290,8 @@ func FirstEverCommit(ctx context.Context, repo api.RepoName) (*gitapi.Commit, er
 	return GetCommit(ctx, repo, id, ResolveRevisionOptions{NoEnsureRevision: true})
 }
 
-// FindNearestCommit finds the commit in the given repository revSpec (e.g. `HEAD` or `mybranch`)
-// whose author date most closely matches the target time.
-//
-// Can return a commit very far away if no nearby one exists.
-// Can theoretically return nil, nil if no commits at all are found.
-func FindNearestCommit(ctx context.Context, repoName api.RepoName, revSpec string, target time.Time) (*gitapi.Commit, error) {
-	if revSpec == "" {
-		revSpec = "HEAD"
-	}
-	// Resolve e.g. the branch we're looking at.
-	branchCommit, err := ResolveRevision(ctx, repoName, revSpec, ResolveRevisionOptions{NoEnsureRevision: true})
-	if err != nil {
-		return nil, err
-	}
-
-	// Find the next closest commit on or after our target time.
-	commitsAfter, err := Commits(ctx, repoName, CommitsOptions{
-		After:     target.Add(-1 * time.Second).Format(time.RFC3339),
-		Range:     string(branchCommit),
-		Reverse:   true,
-		DateOrder: true,
-	})
-	if err != nil {
-		return nil, err
-	}
-	var commitOnOrAfter *gitapi.Commit
-	if len(commitsAfter) > 0 {
-		commitOnOrAfter = commitsAfter[0]
-	}
-
-	// Find the next closest commit before our target time.
-	commitsBefore, err := Commits(ctx, repoName, CommitsOptions{
-		N:         1,
-		Before:    target.Format(time.RFC3339),
-		Range:     string(branchCommit),
-		DateOrder: true,
-	})
-	if err != nil {
-		return nil, err
-	}
-	var commitBefore *gitapi.Commit
-	if len(commitsBefore) > 0 {
-		commitBefore = commitsBefore[0]
-	}
-
-	switch {
-	case commitOnOrAfter == nil && commitBefore == nil:
-		return nil, nil
-	case commitOnOrAfter == nil:
-		return commitBefore, nil
-	case commitBefore == nil:
-		return commitOnOrAfter, nil
-	default:
-		// Get absolute distance of each commit to target.
-		distanceToAfter := commitOnOrAfter.Author.Date.Sub(target)
-		if distanceToAfter < 0 {
-			distanceToAfter = -distanceToAfter
-		}
-		distanceToBefore := commitBefore.Author.Date.Sub(target)
-		if distanceToBefore < 0 {
-			distanceToBefore = -distanceToBefore
-		}
-
-		// Return whichever commit is closer.
-		if distanceToAfter < distanceToBefore {
-			return commitOnOrAfter, nil
-		}
-		return commitBefore, nil
-	}
-}
-
 const (
 	partsPerCommit = 10 // number of \x00-separated fields per commit
-
-	// include refs (slow on repos with many refs)
-	logFormatWithRefs = "--format=format:%H%x00%D%x00%aN%x00%aE%x00%at%x00%cN%x00%cE%x00%ct%x00%B%x00%P%x00"
 
 	// don't include refs (faster, should be used if refs are not needed)
 	logFormatWithoutRefs = "--format=format:%H%x00%x00%aN%x00%aE%x00%at%x00%cN%x00%cE%x00%ct%x00%B%x00%P%x00"
@@ -422,174 +346,4 @@ func parseCommitFromLog(data []byte) (commit *gitapi.Commit, refs []string, rest
 	}
 
 	return commit, refs, rest, nil
-}
-
-// onelineCommit contains (a subset of the) information about a commit returned
-// by `git log --oneline --source`.
-type onelineCommit struct {
-	sha1      string // sha1 commit ID
-	sourceRef string // `git log --source` source ref
-}
-
-// errLogOnelineBatchScannerClosed is returned if a read is attempted on a
-// closed logOnelineBatchScanner.
-var errLogOnelineBatchScannerClosed = errors.New("logOnelineBatchScanner closed")
-
-// logOnelineBatchScanner wraps logOnelineScanner to batch up reads.
-//
-// next will return at least 1 commit and at most maxBatchSize entries. After
-// debounce time it will return what has been batched so far (or wait until an
-// entry is available). The last response from next will return a non-nil
-// error.
-//
-// cleanup must be called when done. This function creates a goroutine to
-// batch up calls to scan.
-func logOnelineBatchScanner(scan func() (*onelineCommit, error), maxBatchSize int, debounce time.Duration) (next func() ([]*onelineCommit, error), cleanup func()) {
-	// done is closed to indicate cleaning up
-	done := make(chan struct{})
-	cleanup = func() {
-		close(done)
-	}
-
-	// We can't call scan with a timeout. So we start a goroutine which sends
-	// the output of scan down a channel. When done is closed, this goroutine
-	// will stop running.
-	type result struct {
-		commit *onelineCommit
-		err    error
-	}
-	// have a cap of maxBatchSize so we potentially have maxBatchSize ready to
-	// send.
-	resultC := make(chan result, maxBatchSize)
-	go func() {
-		defer close(resultC)
-		for {
-			commit, err := scan()
-			select {
-			case resultC <- result{commit: commit, err: err}:
-			case <-done:
-				return
-			}
-			if err != nil {
-				return
-			}
-		}
-	}()
-
-	var err error
-	next = func() ([]*onelineCommit, error) {
-		// The previous call may have had an error but also had commits to
-		// return. In that case we didn't return the error, so need to return
-		// it now.
-		if err != nil {
-			return nil, err
-		}
-
-		// Stop scanning if we haven't found maxBatchSize by deadline.
-		deadline := time.NewTimer(debounce)
-		defer deadline.Stop()
-
-		// We always want 1 result. Ignore deadline for the first result.
-		r, ok := <-resultC
-		if !ok {
-			return nil, errLogOnelineBatchScannerClosed
-		} else if r.err != nil {
-			err = r.err
-			return nil, r.err
-		}
-
-		// Add to commits until we hit maxBatchSize or we are passed
-		// deadline. If we encounter an error set err.
-		commits := []*onelineCommit{r.commit}
-		timedout := false
-		for err == nil && len(commits) < maxBatchSize && !timedout {
-			select {
-			case r, ok := <-resultC:
-				if !ok {
-					err = errLogOnelineBatchScannerClosed
-				} else if r.err != nil {
-					err = r.err
-				} else {
-					commits = append(commits, r.commit)
-				}
-			case <-deadline.C:
-				timedout = true
-			}
-		}
-
-		// Note: err can be non-nil here. The next call to this
-		// function will return err.
-		return commits, nil
-	}
-
-	return next, cleanup
-}
-
-// logOnelineScanner parses the commits from the reader of:
-//
-//   git log --pretty='format:%H %S' -z --source --no-patch
-//
-// Once it returns an error the scanner should be disregarded. io.EOF is
-// returned when there is no more data to read.
-func logOnelineScanner(r io.Reader) func() (*onelineCommit, error) {
-	// Each "log line" contains a source ref. I could not find a bound on the
-	// size of a git ref, so each line can get arbitrarily large. So we use a
-	// bufio.Scanner instead of a bufio.Reader since a Scanner allows growing
-	// the buffer to accomodate the "token" size. This makes the
-	// implementation slightly more complicated (needs a split function
-	// instead of just using ReadBytes).
-	//
-	// Note: Not all source refs correspond to direct arguments, eg if you use
-	// --glob=refs/* any possible ref can be a source ref.
-	//
-	// Note: I check the git source for ref limits, there are none I
-	// found. Linux does have PATH_MAX (4096), but its quite easy to work
-	// around that.
-	//
-	// Note: Scanner does have a max size it will grow to (64kb). If a repo
-	// contains a ref this big, we treat it as an error. This shouldn't happen
-	// in practice, but that is likely famous last words.
-	scanNull := func(data []byte, atEOF bool) (advance int, token []byte, err error) {
-		if atEOF && len(data) == 0 {
-			return 0, nil, nil
-		}
-		if i := bytes.IndexByte(data, '\x00'); i >= 0 {
-			return i + 1, data[:i], nil
-		}
-		if atEOF {
-			return len(data), data, nil
-		}
-		// Request more data.
-		return 0, nil, nil
-	}
-	scanner := bufio.NewScanner(r)
-	scanner.Split(scanNull)
-	return func() (*onelineCommit, error) {
-		if !scanner.Scan() {
-			err := scanner.Err()
-			if err == nil {
-				return nil, io.EOF
-			} else if strings.Contains(err.Error(), "does not have any commits yet") {
-				// Treat empty repositories as having no commits without failing
-				return nil, io.EOF
-			}
-			return nil, err
-		}
-
-		e := scanner.Bytes()
-
-		// Format: (40-char SHA) (source ref)
-		if len(e) <= 42 {
-			return nil, errors.Errorf("parsing git oneline commit: short entry: %q", e)
-		}
-		sha1 := e[:40]
-		if e[40] != ' ' {
-			return nil, errors.Errorf("parsing git oneline commit: no ' ': %q", e)
-		}
-		sourceRef := e[41:]
-		return &onelineCommit{
-			sha1:      string(sha1),
-			sourceRef: string(sourceRef),
-		}, nil
-	}
 }
